+++ upgrade/upgrade-1.6.0.php
@@ -1,0 +1,13 @@
+<?php
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+function upgrade_module_1_6_0($module)
+{
+    return $module instanceof Wilden_manager
+        && $module->registerHook('actionOrderGridDefinitionModifier')
+        && $module->registerHook('actionOrderGridQueryBuilderModifier')
+        && $module->registerHook('displayBackOfficeHeader');
+}